Move DrawPSFStamp to extra_psf.py
